+++ src/handler.js
@@ -4,11 +4,13 @@
  * @license MIT
  * */
 /** IMPORT */
-const Web3 = require('web3')
-const solc = require('solc')
-const { getRemoteCompiler } = require('./compiler/remoteCompiler.js')
-const {readFileCallback} = require('./compiler/utils.js')
 const path = require('path');
+const Web3 = require('web3');
+const solc = require('solc');
+const { getRemoteCompiler } = require('./compiler/remoteCompiler.js');
+const { readFileCallback } = require('./compiler/utils.js');
+const { ExternalProcessBlockchain, ExternalUrlBlockchain ,BuiltinGanacheBlockchain } = require('./blockchain.js');
+
 
 /** CONST */
 const rexTypeErrorReturnArgumentX = /Return argument type (.*) is not implicitly convertible to expected type \(type of first return variable\)/;
@@ -122,6 +124,8 @@ class InteractiveSolidityShell {
             installedSolidityVersion: null, // overridden after merging settings; never use configured value
             providerUrl: 'http://127.0.0.1:8545',
             autostartGanache: true,
+            blockchainProvider: 'internal',
+            ganacheOptions: {},
             ganacheCmd: 'ganache-cli',
             ganacheArgs: [/*'--gasLimit=999000000'*/], //optionally increase default gas limit
             debugShowContract: false,
@@ -139,13 +143,31 @@ class InteractiveSolidityShell {
 
         this.cache = {
             compiler: {} /** compilerVersion:object */
-        }
+        };
 
         this.cache.compiler[this.settings.installedSolidityVersion.startsWith("^") ? this.settings.installedSolidityVersion.substring(1) : this.settings.installedSolidityVersion] = solc;
-        this.reset()
+        this.reset();
 
-        this.blockchain = new Blockchain(this)
-        this.blockchain.connect()
+        this.initBlockchain();
+    }
+
+    initBlockchain() {
+        if(this.blockchain){
+            this.blockchain.stopService();
+        }
+
+        if(!this.settings.blockchainProvider || this.settings.blockchainProvider === "internal"){
+            this.blockchain = new BuiltinGanacheBlockchain(this);
+        } else if(this.settings.blockchainProvider.startsWith("https://") || this.settings.blockchainProvider.startsWith("http://")) {
+            this.blockchain = new ExternalUrlBlockchain(this, this.settings.blockchainProvider);
+        } else if (this.settings.blockchainProvider.length > 0) {
+            this.settings.ganacheCmd = this.settings.blockchainProvider;
+            this.blockchain = new ExternalProcessBlockchain(this);
+        } else {
+            this.log("  🧨 unknown blockchain provider. falling back to built-in ganache.")
+            this.blockchain = new BuiltinGanacheBlockchain(this);
+        }
+        this.blockchain.connect();
     }
 
     loadSession(stmts) {
@@ -407,113 +429,6 @@ contract ${this.settings.templateContractName} {
                 })
             })
         });
-    }
-}
-
-class Blockchain {
-    constructor(shell) {
-        this.log = shell.log;
-        this.shell = shell
-
-        this.provider = undefined
-        this.web3 = undefined
-        this.deployed = {}
-
-        this.proc;
-    }
-
-    connect() {
-        this.provider = new Web3.providers.HttpProvider(this.shell.settings.providerUrl);
-        this.web3 = new Web3(this.provider);
-
-        this.web3.eth.net.isListening().then().catch(err => {
-            if (!this.shell.settings.autostartGanache) {
-                console.warn("⚠️  ganache autostart is disabled")
-                return;
-            }
-            this.startService()
-            this.provider = new Web3.providers.HttpProvider(this.shell.settings.providerUrl);
-            this.web3 = new Web3(this.provider);
-        })
-
-    }
-
-    startService() {
-        if (this.proc) {
-            return this.proc;
-        }
-        this.log("ℹ️  ganache-mgr: starting temp. ganache instance ...\n »");
-        
-        this.proc = require('child_process').spawn(this.shell.settings.ganacheCmd, this.shell.settings.ganacheArgs);
-        this.proc.on('error', function(err) {
-            console.error(`
- 🧨 Unable to launch blockchain serivce: ➜ ℹ️  ${err}
-
-    Please verify that 'ganache-cli' (or similar service) is installed and available in your PATH.
-    Otherwise, you can disable autostart by setting 'autostartGanache' to false in your settings or configure a different service and '.restartblockchain'.
-    `);
-          });
-        
-        
-    }
-
-    stopService() {
-        this.log("💀  ganache-mgr: stopping temp. ganache instance");
-        if(this.proc) {
-            this.proc.kill('SIGINT');
-            this.proc = undefined;
-        }
-    }
-
-    restartService() {
-        this.stopService();
-        this.startService();
-    }
-
-    getAccounts() {
-        return new Promise((resolve, reject) => {
-            this.web3.eth.getAccounts((err, result) => {
-                if (err) return reject(new Error(err));
-                return resolve(result);
-            })
-        });
-    }
-
-    async deploy(contracts, callback) {
-        //sort deploy other contracts first
-        Object.entries(contracts).sort((a, b) => a[1].main ? 10 : -1).forEach(([templateContractName, o]) => {
-            if (o.evm.bytecode.object.length === 0) {
-                return; //no bytecode, probably an interface
-            }
-
-            let thisContract = {
-                bytecode: o.evm.bytecode.object,
-                abi: o.abi,
-                proxy: new this.web3.eth.Contract(o.abi, null),
-                instance: undefined,
-                main: o.main,
-                accounts: undefined
-            }
-
-            this.deployed[templateContractName] = thisContract;
-            this.getAccounts()
-                .then(accounts => {
-                    thisContract.accounts = accounts;
-                    let instance = thisContract.proxy.deploy({ data: thisContract.bytecode }).send({ from: accounts[0], gas: this.shell.settings.deployGas })
-                    thisContract.instance = instance;
-                    return instance;
-                })
-                .then(contract => {
-                    if (thisContract.main) {
-                        contract.methods[thisContract.main]().call({ from: thisContract.accounts[0], gas: this.shell.settings.callGas }, callback);
-                    }
-                    return;
-                })
-                .catch(err => {
-                    callback(`💥  ganache not yet ready. Please try again. (👉 ${err} 👈)`)
-                })
-
-        }, this);
     }
 }